+++ gradio/chat_interface.py
@@ -198,12 +198,32 @@ class ChatInterface(Blocks):
                 )
 
             self.saved_input = State()
+            self.chatbot_state = State([])
 
             self._setup_events()
             self._setup_api()
 
     def _setup_events(self) -> None:
         submit_fn = self._stream_fn if self.is_generator else self._submit_fn
+
+        self.textbox.submit(
+            self._clear_and_save_textbox,
+            [self.textbox],
+            [self.textbox, self.saved_input],
+            api_name=False,
+            queue=False,
+        ).then(
+            self._display_input,
+            [self.saved_input, self.chatbot_state],
+            [self.chatbot, self.chatbot_state],
+            api_name=False,
+            queue=False,
+        ).then(
+            submit_fn,
+            [self.saved_input, self.chatbot_state],
+            [self.chatbot, self.chatbot_state],
+            api_name=False,
+        )
 
         submit_event = (
             self.textbox.submit(
@@ -215,15 +235,15 @@ class ChatInterface(Blocks):
             )
             .then(
                 self._display_input,
-                [self.saved_input, self.chatbot],
-                [self.chatbot],
+                [self.saved_input, self.chatbot_state],
+                [self.chatbot, self.chatbot_state],
                 api_name=False,
                 queue=False,
             )
             .then(
                 submit_fn,
-                [self.saved_input, self.chatbot],
-                [self.chatbot],
+                [self.saved_input, self.chatbot_state],
+                [self.chatbot, self.chatbot_state],
                 api_name=False,
             )
         )
@@ -282,7 +302,7 @@ class ChatInterface(Blocks):
         if self.undo_btn:
             self.undo_btn.click(
                 self._delete_prev_fn,
-                [self.chatbot],
+                [self.chatbot_state],
                 [self.chatbot, self.saved_input],
                 api_name=False,
                 queue=False,
@@ -296,9 +316,9 @@ class ChatInterface(Blocks):
 
         if self.clear_btn:
             self.clear_btn.click(
-                lambda: ([], None),
+                lambda: ([], [], None),
                 None,
-                [self.chatbot, self.saved_input],
+                [self.chatbot, self.chatbot_state, self.saved_input],
                 queue=False,
                 api_name=False,
             )
@@ -348,14 +368,10 @@ class ChatInterface(Blocks):
     def _setup_api(self) -> None:
         api_fn = self._api_stream_fn if self.is_generator else self._api_submit_fn
 
-        # Use a gr.State() instead of self.chatbot so that the API doesn't require passing forth
-        # a chat history, instead it is just stored internally in the state.
-        history = State([])
-
         self.fake_api_btn.click(
             api_fn,
-            [self.textbox, history],
-            [self.textbox, history],
+            [self.textbox, self.chatbot_state],
+            [self.textbox, self.chatbot_state],
             api_name="chat",
         )
 
@@ -364,30 +380,33 @@ class ChatInterface(Blocks):
 
     def _display_input(
         self, message: str, history: list[list[str | None]]
-    ) -> list[list[str | None]]:
+    ) -> tuple[list[list[str | None]], list[list[str | None]]]:
         history.append([message, None])
-        return history
+        return history, history
 
     def _submit_fn(
         self, message: str, history_with_input: list[list[str | None]]
-    ) -> list[list[str | None]]:
+    ) -> tuple[list[list[str | None]], list[list[str | None]]]:
         history = history_with_input[:-1]
         response = self.fn(message, history)
         history.append([message, response])
-        return history
+        return history, history
 
     def _stream_fn(
         self, message: str, history_with_input: list[list[str | None]]
-    ) -> Generator[list[list[str | None]], None, None]:
+    ) -> Generator[tuple[list[list[str | None]], list[list[str | None]]], None, None]:
         history = history_with_input[:-1]
         generator = self.fn(message, history)
         try:
             first_response = next(generator)
-            yield history + [[message, first_response]]
+            update = history + [[message, first_response]]
+            yield update, update
         except StopIteration:
-            yield history + [[message, None]]
+            update = history + [[message, None]]
+            yield update, update
         for response in generator:
-            yield history + [[message, response]]
+            update = history + [[message, response]]
+            yield update, update
 
     def _api_submit_fn(
         self, message: str, history: list[list[str | None]]
@@ -419,9 +438,9 @@ class ChatInterface(Blocks):
 
     def _delete_prev_fn(
         self, history: list[list[str | None]]
-    ) -> tuple[list[list[str | None]], str]:
+    ) -> tuple[list[list[str | None]], str, list[list[str | None]]]:
         try:
             message, _ = history.pop()
         except IndexError:
             message = ""
-        return history, message or ""
+        return history, message or "", history